refactor(toolkit-history): extract shared history-format helpers
formatDuration and getInitials were byte-identical copies duplicated
across HistoryTable, HistoryCards, HistoryDetailModal (client.tsx also
duplicated formatDuration), and MAX_AVATARS was duplicated in
HistoryTable/HistoryCards. Extract all three into a pure _lib module
with dedicated unit tests. Pure refactor, no behavior change.
