+++ apps/web/src/app/(authenticated)/toolkit/history/_components/HistoryDetailModal.tsx
@@ -13,28 +13,13 @@ import {
 import { useTranslation } from '@/hooks/useTranslation';
 import { cn } from '@/lib/utils';
 
+import { formatDuration, getInitials } from '../_lib/history-format';
+
 import type { HistoryRow } from '../_lib/history-filters';
 
 export interface HistoryDetailModalProps {
   row: HistoryRow | null;
   onClose: () => void;
-}
-
-/** Formats a session duration in minutes as `"{h}h {m}m"` — mirrors HistoryTable/HistoryCards. */
-function formatDuration(minutes: number): string {
-  const h = Math.floor(minutes / 60);
-  const m = minutes % 60;
-  return `${h}h ${m}m`;
-}
-
-/** Derives up to 2 uppercase initials from a player's display name — mirrors HistoryTable/HistoryCards. */
-function getInitials(name: string): string {
-  const parts = name.trim().split(/\s+/).filter(Boolean);
-  if (parts.length === 0) return '?';
-  if (parts.length === 1) return (parts[0] as string).slice(0, 2).toUpperCase();
-  const first = parts[0] as string;
-  const last = parts[parts.length - 1] as string;
-  return `${first[0]}${last[0]}`.toUpperCase();
 }
 
 /**
